refactor(SpanRow): Move SpanRow component to SpanRow/SpanItemRow and update import path
