--- src/Routes/Tables/Jobs/Trace/SpanRow.jsx
@@ -1,3 +1,0 @@
-import SpanRow from './SpanRow/index';
-
-export default SpanRow;
